Feat: MyPage BE

+++ src/main/java/com/ureca/myspring/repository/MemberRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ureca.myspring.dto.MemberDTO;
 
 
-public interface MemberRepository extends JpaRepository<MemberDTO, Integer>{
+public interface MemberRepository extends JpaRepository<MemberDTO, Long>{
 }
